scripts/vita/parse_core/requirements.txt
pyelftools==0.24
